+++ DifferentiationInterfaceTest/test/zero_backends.jl
@@ -7,6 +7,11 @@ using DifferentiationInterfaceTest
 using DifferentiationInterfaceTest: allocfree_scenarios, no_matrices
 using Test
 
+## Counterparts
+
+test_counterparts(AutoZeroForward())
+test_counterparts(AutoZeroReverse())
+
 ## Type stability
 
 test_differentiation(